Cap 58, 59 y 60

+++ app/Http/Controllers/Api/ArticleController.php
@@ -3,19 +3,22 @@
 namespace App\Http\Controllers\Api;
 
 use App\Models\Article;
-use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\ArticleResource;
 use App\Http\Requests\SaveArticleRequest;
-use App\Http\Resources\ArticleCollection;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Http\Resources\Json\JsonResource;
 use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+
+
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class ArticleController extends Controller implements \Illuminate\Routing\Controllers\HasMiddleware
 {
+    use AuthorizesRequests;
 
     public static function middleware(): array
     {
@@ -47,23 +50,26 @@ class ArticleController extends Controller implements \Illuminate\Routing\Contro
 
     public function store(SaveArticleRequest $request){
 
-        
+        $this->authorize('create', new Article);
+
         $article = Article::create($request->validated());
 
         return ArticleResource::make($article);
     }
 
-    public function update(Article $article, SaveArticleRequest $request){
-
-     
-
+    public function update(Article $article, SaveArticleRequest $request): ArticleResource
+    {
+        $this->authorize('update', $article);
         $article->update($request->validated());
 
         return ArticleResource::make($article);
     }
 
-    public function destroy(Article $article): Response
+    public function destroy(Article $article, Request $request): Response
     {   
+        
+        $this->authorize('delete', $article);
+
         $article->delete();
         return response()->noContent();
